Refactor Dex component to use BeastsDexGrid and add associated styles
