+++ client/src/main.tsx
@@ -16,7 +16,7 @@ import NewCover from "./components/NewCover/index.tsx";
 import Tamagotchi from "./components/Tamagotchi/index.tsx";
 import Bag from "./components/Bag/index.tsx";
 import Chat from "./components/Chat/index.tsx";
-import Dex from "./components/Dex/Grid.tsx";
+import Dex from "./components/BeastsDexGrid/index.tsx";
 
 // Import the layout component
 import AppLayout from "./components/Layouts/AppLayout.tsx";
